Consolidate and standardize sysimage tools
